[Refactor] Move files to record_detail directory
