+++ app/src/main/res/layout/fragment_record_detail_list.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/white"
-    tools:context=".ui.record.RecordDetailListFragment">
+    tools:context=".ui.record_detail.RecordDetailListFragment">
 
     <com.naver.maps.map.MapView
         android:id="@+id/map_record_detail_list"
