+++ app/src/main/res/layout/fragment_record_detail_map.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/white"
-    tools:context=".ui.record.RecordDetailMapFragment">
+    tools:context=".ui.record_detail.RecordDetailMapFragment">
 
     <com.naver.maps.map.MapView
         android:id="@+id/mv_record_detail_full_map"
